got basic tests to run, need to figure out why repository queries aren't defined.

+++ src/users/users.controller.ts
@@ -1,7 +1,9 @@
-import { Controller, Get, Post, Body, Patch, Param, Delete } from '@nestjs/common';
+import { Controller, Get, Post, Body, Patch, Param, Delete, Res, BadRequestException } from '@nestjs/common';
 import { UsersService } from './users.service';
 import { CreateUserDto } from './dto/create-user.dto';
 import { UpdateUserDto } from './dto/update-user.dto';
+import { isUUID } from 'class-validator';
+import { Response } from 'express';
 
 @Controller('users')
 export class UsersController {
@@ -19,16 +21,25 @@ export class UsersController {
 
   @Get(':id')
   findOne(@Param('id') id: string) {
+    if (!isUUID(id)) {
+      throw new BadRequestException()
+    }
     return this.usersService.findOne(id);
   }
 
   @Patch(':id')
   update(@Param('id') id: string, @Body() updateUserDto: UpdateUserDto) {
+    if (!isUUID(id)) {
+      throw new BadRequestException()
+    }
     return this.usersService.update(id, updateUserDto);
   }
 
   @Delete(':id')
   remove(@Param('id') id: string) {
+    if (!isUUID(id)) {
+      throw new BadRequestException()
+    }
     return this.usersService.delete(id);
   }
 }
